services/notifier/requirements.txt
python-telegram-bot==20.7
pika==1.3.2
psycopg2-binary==2.9.9
structlog==23.2.0
python-dotenv==1.0.0
